Database: Start parsing setup.ini into database item
Here is a great process. Within the cycle of yylex(), the parser's
routine can be formulated word-by-word, but to all intents it's not an
easy job.

As yylex() parses a setup.ini file word-by-word, having markers marking
out if I'm handling a package info / YAML item, should be the key.