lexer: FIX: Package name could include "."
"." is also allowed in a package name. Usually used in distinguishing
different explicit versions. For example:

@ autoconf2.1 - Stable version
@ autoconf2.5 - Current version

What's the difference: Autoconf2.1 contains autoconf-2.13, and is meant
to be installed alongside autoconf2.5 (which contains autoconf-2.5x)
and alongside the 'autoconf' package, which contains wrapper scripts
which call the appropriate 'real' autoconf, stable or devel, based on
target package heuristics.