database: No need to sqlite3_free(zErrMsg)
sqlite3_free(zErrMsg) is to free the variable zErrMsg containing error
message. This call is used in Runoob's example, because the author will
drop the message after showing it.

But in Cygpm, zErrMsg is defined as a class member so that other relatives
can read it after a sqlite3_exec(). No need to free it so fast.