database: main: Remove database/journal before building database
Database should always be regenerated when update, so removing it at
first instead of dropping tables will save time making a large journal
as backup.

Also, to collaborate with DB Browser, if a database is occupied (usually
by DB Browser), database will not be removed, then fall back to "DROP
TABLE IF EXISTS".