database: insert(Prev)PackageInfo: Prevent illegal chars
By default, char array may contain illegal characters when not
initialized. These illegal chars will not be recognized by SQLite, and
SQLite will throw an error.

This is happened when parsing source__raw. Not every package have
source__raw, so when sscanf() parsing an empty source__raw, char arrays
receiving results will keep as-is, not being overwritten by legal
characters. In consequence, a package will not be added unexpectedly.

To prevent this bug, we must initialize those char arrays with an empty
string.